+++ Arduino.h
@@ -9,6 +9,6 @@
 void pinMode(unsigned char pin, unsigned char mode);
 void digitalWrite(unsigned char pin, unsigned char state);
 void delayMicroseconds(int millis);
-int pulseIn(unsigned char pin, unsigned char state, int timeout);
+unsigned long pulseIn(unsigned char pin, unsigned char state, int timeout);
 
 #endif /* __ARDUINO_PARKING_SENSOR_ARDUINO__ */
